# docs/api/openapi-dar.yaml
openapi: 3.1.0
info:
  title: Data Access Register
  description: "Core APIs for the Central Data Access Register.\n\nThe register records\
    \ all lawful access to customer energy data by Controllers,\nregardless of the\
    \ legal basis on which that access rests. It is not exclusively\na consent register\
    \ — legitimate interests, public task, legal obligation, and\ncontract are all\
    \ within scope alongside consent-based access.\n\n**Jurisdiction:** This is a\
    \ UK-only solution. All Controllers, customers, and\nmeter points are within Great\
    \ Britain. The governing data protection framework\nis **UK GDPR** and the **Data\
    \ Protection Act 2018**. EU GDPR does not apply.\nThe jurisdiction field is fixed\
    \ to `GB` and does not need to be supplied.\n\n## Purpose\n\nThe register provides\
    \ a single, auditable source of truth for:\n\n- **Controllers** — to record and\
    \ demonstrate the legal basis under which they\n  access customer meter data.\n\
    - **Data Providers** — to verify that a Controller has a registered, active access\n\
    \  record before releasing data.\n- **Customers** — to view all parties registered\
    \ as having lawful access to their\n  meter point, on request via any Controller.\n\
    \n## Party Model\n\n| Party | UK GDPR Role | Description |\n|---|---|---|\n| **Data\
    \ User** | Processor | Operates the register. Fetches and processes meter data\
...
    \ Current version: `v1`.\n"
  version: 0.0.13
...
servers:
...
paths:
...
  /access-records:
    post:
      operationId: createAccessRecord
      summary: Create an Access Record
      description: "Registers a new access record on behalf of a Controller. Returns\
        \ `201` and\nan `access-token` containing the new access key (`ak`), plus\
        \ a `Location`\nheader pointing to the created resource.\n\nAll records use\
        \ the same `AccessRecord` schema. The `processing.legal-basis`\nvalue determines\
        \ which additional fields are expected:\n\n- **Consent bases** — populate\
        \ `notice` and `access-event.consent`.\n- **Legitimate interests** — populate\
        \ `lia-reference` on the lead controller\n  entry; set `notice` and `access-event.consent`\
        \ to null.\n- **Public task / legal obligation** — populate `statutory-reference`\
        \ on the\n  lead controller entry; set `notice` and `access-event.consent`\
        \ to null.\n- **Contract** — set `notice` and `access-event.consent` to null.\n\
        \nHistoric records may be submitted — `access-event.registered-at` may predate\n\
        `record-metadata.created-at`.\n\nIf the DCC has previously submitted a `DISCOVERED`\
        \ record for the same MPxN\nand organisation, the register will link the two\
        \ records on creation. The\nDISCOVERED record is retained for audit; the new\
        \ full record becomes `ACTIVE`.\n\nTo update an existing record use `PUT /access-records/{ak}`.\n\
        \n\n**Cross-DUID re-identification:** To create an Access Record linked to\
        \ an Identity Record originally created by another Data User, include a `reidentification-token`\
        \ field in the request body (alongside the standard `AccessRecord` fields,\
        \ without `record-metadata.identity-record-ref`). Obtain the token via `POST\
        \ /identity-records/reidentify`. The register validates the token and resolves\
        \ the `ir` internally."
      tags:
      - Data Users
      security:
      - bearerAuth: []
      requestBody:
        description: An AccessRecord. All legal bases use this same schema.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRecord'
            examples:
              consent:
                summary: Consent-based access — sole controller
                value:
                  record-metadata:
                    schema-version: '1.0'
                    controller-arrangement:
                      arrangement-type: sole
                      controllers:
                      - name: Bright Energy Ltd
                        role: sole
                        contact-url: https://bright-energy.com/contact
                        address:
                          addressLine1: 10 Canary Wharf
                          townCity: London
                          postcode: E14 5AB
                        privacy-rights-url: https://bright-energy.com/your-rights
                    identity-record-ref: ir_a3c5e7f9b1d3a3c5e7f9b1d3
                  notice:
                    shared-notice:
                      terms-url: https://bright-energy.com/privacy-v3.html
                      notice-version: v3.2
                      notice-language: en
                  processing:
                    legal-basis: uk-consent
                    purpose: Energy efficiency analysis and tariff recommendations
                    data-types:
                    - HH-CONSUMPTION
                    - HH-EXPORT
                  access-event:
                    state: ACTIVE
                    registered-at: '2023-11-07T05:31:56Z'
                    expiry: '2027-11-10T17:07:01.580Z'
                    controller-reference: REF-00123
                    consent:
                      consent-type: expressed-consent
                      method: Explicit web checkbox
              consent-joint:
                summary: Consent-based access — joint controllers (Art. 26)
                value:
                  record-metadata:
                    schema-version: '1.0'
                    controller-arrangement:
                      arrangement-type: joint
                      art26-reference: JCA-2024-BrightGrid-001
                      controllers:
                      - name: Bright Energy Ltd
                        role: lead
                        contact-url: https://bright-energy.com/contact
                        address:
                          addressLine1: 10 Canary Wharf
                          townCity: London
                          postcode: E14 5AB
                        privacy-rights-url: https://bright-energy.com/your-rights
                      - name: Grid Analytics Ltd
                        role: joint
                        contact-url: https://gridanalytics.com/contact
                        privacy-rights-url: https://gridanalytics.com/your-rights
                        lia-reference: LIA-2024-003
                    identity-record-ref: ir_a3c5e7f9b1d3a3c5e7f9b1d3
                  notice:
                    notices:
                    - controller-name: Bright Energy Ltd
                      terms-url: https://bright-energy.com/privacy-v3.html
                      notice-version: v3.2
                      notice-language: en
                    - controller-name: Grid Analytics Ltd
                      terms-url: https://gridanalytics.com/privacy-v2.html
                      notice-version: v2.0
                      notice-language: en
                  processing:
                    legal-basis: uk-consent
                    purpose: Energy efficiency analysis and demand forecasting
                    data-types:
                    - HH-CONSUMPTION
                    - HH-EXPORT
                  access-event:
                    state: ACTIVE
                    registered-at: '2024-03-01T09:00:00Z'
                    expiry: '2027-03-01T00:00:00Z'
                    controller-reference: REF-00456
                    consent:
                      consent-type: expressed-consent
                      method: Explicit web checkbox
              consent-group:
                summary: Consent-based access — group (trading names)
                value:
                  record-metadata:
                    schema-version: '1.0'
                    controller-arrangement:
                      arrangement-type: group
                      controllers:
                      - name: Bright Energy Group plc
                        role: lead
                        contact-url: https://bright-energy.com/contact
                        privacy-rights-url: https://bright-energy.com/your-rights
                      - name: Bright Energy Home Services
                        role: member
                        contact-url: https://bright-energy.com/home/contact
                      - name: Bright Energy Smart
                        role: member
                        contact-url: https://bright-energy.com/smart/contact
                    identity-record-ref: ir_a3c5e7f9b1d3a3c5e7f9b1d3
                  notice:
                    shared-notice:
                      terms-url: https://bright-energy.com/group-privacy-v1.html
                      notice-version: grp-v1.0
                      notice-language: en
                  processing:
                    legal-basis: uk-consent
                    purpose: Energy efficiency analysis and tariff recommendations
                    data-types:
                    - HH-CONSUMPTION
                  access-event:
                    state: ACTIVE
                    registered-at: '2024-06-01T10:00:00Z'
                    expiry: '2028-06-01T00:00:00Z'
                    controller-reference: REF-GRP-001
                    consent:
                      consent-type: expressed-consent
                      method: Explicit web checkbox
              legitimate-interests:
                summary: Legitimate interests access
                value:
                  record-metadata:
                    schema-version: '1.0'
                    controller-arrangement:
                      arrangement-type: sole
                      controllers:
                      - name: Grid Analytics Ltd
                        role: sole
                        contact-url: https://gridanalytics.com/contact
                        address:
                          addressLine1: 5 Tech Park
                          townCity: Manchester
                          postcode: M1 2AB
                        privacy-rights-url: https://gridanalytics.com/your-rights
                        lia-reference: LIA-2024-003
                    identity-record-ref: ir_b4d6f8a0c2e4b4d6f8a0c2e4
                  notice: null
                  processing:
                    legal-basis: uk-legitimate-interests
                    purpose: Network load balancing and demand forecasting
                    data-types:
                    - HH-CONSUMPTION
                  access-event:
                    state: ACTIVE
                    registered-at: '2024-03-01T09:00:00Z'
                    expiry: '2026-03-01T00:00:00Z'
                    controller-reference: CONTRACT-9981
                    consent: null
              public-task:
                summary: Public task access (regulated/statutory body)
                value:
                  record-metadata:
                    schema-version: '1.0'
                    controller-arrangement:
                      arrangement-type: sole
                      controllers:
                      - name: Northern Networks plc
                        role: sole
                        contact-url: https://northernnetworks.co.uk/contact
                        address:
                          addressLine1: 1 Grid House
                          townCity: Leeds
                          postcode: LS1 1AA
                        privacy-rights-url: https://northernnetworks.co.uk/your-rights
                        statutory-reference: Energy Act 2023, s.147
                    identity-record-ref: ir_c5e7a9b1d3f5c5e7a9b1d3f5
                  notice: null
                  processing:
                    legal-basis: uk-public-task
                    purpose: Statutory network safety and capacity monitoring
                    data-types:
                    - HH-CONSUMPTION
                    - HH-EXPORT
                  access-event:
                    state: ACTIVE
                    registered-at: '2024-01-15T00:00:00Z'
                    expiry: null
                    controller-reference: OFGEM-AUTH-2024-0042
                    consent: null
      responses:
        '201':
          description: Access record created successfully.
          headers:
            Location:
              description: URL of the newly created access record.
              schema:
                type: string
                example: /v1/access-records/ak_691df0c788ca043403b7fa90
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /access-records/{ak}:
    get:
      operationId: verifyAccess
      summary: Verify Access Record
      description: "Returns the full access record for a given access key. Called\
        \ by Data Providers\nbefore releasing meter data.\n\n**Authentication:** Unauthenticated\
        \ by design. The access key acts as the\ncredential; treat it as a secret.\n\
        \n**Note on PII:** This endpoint returns no PII. The person-property relationship\n\
        (MPxN, address, move-in date, identity verification) is held in the linked\n\
        Identity Record, which is accessible only to authenticated Data Users via\n\
        `GET /identity-records/{ir}`. The `identity-record-ref` in `record-metadata`\n\
        provides the link.\n\n**Verification checklist (all record types):**\n1. `access-event.state`\
        \ must be `ACTIVE`.\n2. `access-event.expiry` must be in the future, or null.\n\
        3. At least one controller in `record-metadata.controller-arrangement.controllers`\n\
        \   must match the organisation you have a verified commercial relationship\
        \ with\n   on your platform. For `joint` arrangements, verify against the\
        \ `lead` entry.\n4. The data types being requested must be within `processing.data-types`.\n\
        5. Log `response.transaction-id` for your audit trail.\n\n**Additional checks\
        \ by legal basis:**\n- Consent records: confirm `access-event.consent.consent-type`\
        \ is appropriate\n  for the data category being released.\n- Legitimate interests:\
        \ confirm `lia-reference` is present on the lead\n  controller entry in `record-metadata.controller-arrangement`.\n\
        - Public task / legal obligation: confirm `statutory-reference` is present\n\
        \  on the lead controller entry in `record-metadata.controller-arrangement`.\n"
      tags:
      - Data Providers
      parameters:
      - name: ak
        in: path
        description: The access key issued at registration.
        required: true
        schema:
          $ref: '#/components/schemas/AccessKey'
      responses:
        '200':
          description: Access record found and returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyAccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      operationId: replaceAccessRecord
      summary: Replace an Access Record
      description: "Replaces the full payload of an existing access record. The record\
        \ identified\nby `ak` must have been registered by the authenticated Data\
        \ User.\n\nThis is a full replacement — the entire `AccessRecord` body must\
        \ be supplied.\nThe `ak`, `duid`, `created-at`, and `record-metadata.record-identifier`\
        \ are\npreserved from the original; all other fields are replaced with the\
        \ submitted\npayload.\n\nUse this to:\n- Refresh a consent record when the\
        \ customer renews (e.g. update `expiry`\n  or `notice-version`). To update\
        \ identity verification evidence, update\n  the linked Identity Record via\
        \ `POST /identity-records` and update\n  `record-metadata.identity-record-ref`\
        \ to the new `ir` key.\n- Correct an error in a submitted record\n- Update\
        \ `processing.purpose` or `data-types` after a scope change\n\n**State constraint:**\
        \ Only `ACTIVE` records may be replaced. Attempting to\nreplace a `REVOKED`\
        \ or `EXPIRED` record returns `409`.\n"
      tags:
      - Data Users
      security:
      - bearerAuth: []
      parameters:
      - name: ak
        in: path
        description: The access key of the record to replace.
        required: true
        schema:
          $ref: '#/components/schemas/AccessKey'
      requestBody:
        description: Full AccessRecord replacement payload.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRecord'
      responses:
        '200':
          description: Access record replaced successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterAccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 'Record is not in `ACTIVE` state and cannot be replaced. `REVOKED`
            and `EXPIRED` records are immutable.

            '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: revokeAccess
      summary: Revoke Access Record
      description: 'Revokes an access record. Transitions the record to `REVOKED`
        state with a

        `revoked-at` timestamp. Records are retained for audit — not deleted.


        For consent records this represents customer withdrawal of consent.

        For non-consent records this represents the Controller removing their

        access registration (e.g. LIA no longer applicable, contract ended,

        statutory authority lapsed).


        After revocation `GET /access-records/{ak}` returns the record in `REVOKED`

        state. Data Providers must deny access for any non-`ACTIVE` record.

        '
      tags:
      - Data Users
      security:
      - bearerAuth: []
      parameters:
      - name: ak
        in: path
        description: The access key to revoke.
        required: true
        schema:
          $ref: '#/components/schemas/AccessKey'
      responses:
        '200':
          description: Access record revoked. Record retained in REVOKED state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeAccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /meter-points/{mpxn}/access-records:
    get:
      operationId: listAccessRecords
      summary: List Access Records for a Meter Point
      description: '<Warning>Could be fully disabled if [Central Portal](/dar/customer-portal)
        is prefered mechanism</Warning>


        Returns access records associated with a given MPxN, registered by the

        authenticated Data User. Covers all legal bases and all lifecycle states

        unless filtered. Supports customer transparency — customers may request

        this list via any Controller to see who holds registered access to their

        meter point data.


        `DISCOVERED` records submitted by the DCC are also returned in this

        response, allowing customers to see organisations known to have accessed

        their meter data historically that have not yet registered as Data Users.

        DISCOVERED records are clearly distinguished by their `state` field and

        contain only the sparse `discovered-access` object rather than a full

        access record.

        '
      tags:
      - Data Users
      security:
      - bearerAuth: []
      parameters:
      - name: mpxn
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/MPxN'
      - name: state
        in: query
        description: Filter by record lifecycle state. Omit to return all states.
        required: false
        schema:
          $ref: '#/components/schemas/AccessState'
      - name: legal-basis
        in: query
        description: Filter by legal basis. Omit to return all bases.
        required: false
        schema:
          $ref: '#/components/schemas/LegalBasis'
      responses:
        '200':
          description: Access record list returned successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAccessRecordsResponse'
        '401':
          $ref: '#/components/responses/Unauthorised'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Base64-encode `account_id:secret_key`. Used only on `GET /auth/token`.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT from `GET /auth/token`. Pass as `Authorization: Bearer <token>`.
        Expires after 7200s.'
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorised:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated caller lacks permission for this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: An unexpected server-side error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    MPxN:
      type: string
      pattern: ^(?:[0-9A-HJ-NPR-Z]{2}[0-9]{8,10}|[0-9]{10})$
      description: Meter Point Administration Number (MPAN) or Meter Point Reference
        Number (MPRN).
      example: '1234567890123'
    AccessKey:
      type: string
      pattern: ^ak_[0-9a-f]{24}$
      description: 'Unique opaque identifier for an access record, issued by the register
        on creation. Treat as a secret — possession enables access verification.

        '
      example: ak_691df0c788ca043403b7fa90
...
    AccessState:
      type: string
      description: "Lifecycle state of the access record.\n\n- `ACTIVE` — access is\
        \ registered and in force.\n- `EXPIRED` — the record's `expiry` has passed.\n\
        - `REVOKED` — the Controller has removed their registration, or the customer\n\
        \  has withdrawn consent. Retained for audit with `access-event.revoked-at`\
        \ set.\n- `DISCOVERED` — a sparse record created by the DCC from historic\
        \ meter\n  transaction logs. The organisation has not yet registered as a\
        \ Data User.\n  Visible to customers for transparency but cannot authorise\
        \ data release.\n  Transitions to superseded when the organisation registers\
        \ a full record.\n"
      enum:
      - ACTIVE
      - EXPIRED
      - REVOKED
      - DISCOVERED
    LegalBasis:
      type: string
      description: 'The legal basis under which the Controller accesses customer data.


        **Consent bases** — `notice` and `access-event.consent` must be populated.


        | Value | Plain name | Article |

        |---|---|---|

        | `uk-consent` | UK Consent | UK GDPR Art. 6(1)(a) |

        | `uk-explicit-consent` | UK Explicit Consent | UK GDPR Art. 9(2)(a) |


        **Non-consent bases** — `notice` and `access-event.consent` must be null.


        | Value | Plain name | Article | Supporting field (on lead controller) |

        |---|---|---|---|

        | `uk-legitimate-interests` | UK Legitimate Interests | UK GDPR Art. 6(1)(f)
        | `lia-reference` |

        | `uk-public-task` | UK Public Task | UK GDPR Art. 6(1)(e) | `statutory-reference`
        |

        | `uk-legal-obligation` | UK Legal Obligation | UK GDPR Art. 6(1)(c) | `statutory-reference`
        |

        | `uk-contract` | UK Contract | UK GDPR Art. 6(1)(b) | — |

        '
      enum:
      - uk-consent
      - uk-explicit-consent
      - uk-legitimate-interests
      - uk-public-task
      - uk-legal-obligation
      - uk-contract
      example: uk-consent
...
    AccessRecord:
      type: object
      description: 'A data access record covering all lawful bases. The same schema
        is used

        regardless of `processing.legal-basis`. The `notice` field and

        `access-event.consent` sub-object are null for non-consent records and

        populated for consent records.

        '
      properties:
        record-metadata:
          $ref: '#/components/schemas/RecordMetadata'
        notice:
          nullable: true
          description: 'Populated for consent-based records; null for all other legal
            bases.

            '
          allOf:
          - $ref: '#/components/schemas/NoticeFields'
        processing:
          $ref: '#/components/schemas/ProcessingFields'
        access-event:
          $ref: '#/components/schemas/AccessEventFields'
      required:
      - record-metadata
      - notice
      - processing
      - access-event
    AccessToken:
      type: object
      description: 'The access token issued on successful registration. Pass to Data
        Providers to enable access verification.

        '
      properties:
        key:
          $ref: '#/components/schemas/AccessKey'
        expiry:
          type: string
          format: date-time
          nullable: true
          description: Mirrors `access-event.expiry` for convenience.
          example: '2027-11-10T17:07:01.580Z'
      required:
      - key
    RegisterAccessResponse:
      type: object
      properties:
        response:
          $ref: '#/components/schemas/ResponseMetadata'
        access-token:
          $ref: '#/components/schemas/AccessToken'
      required:
      - response
      - access-token
    VerifyAccessResponse:
      type: object
      description: 'The full access record returned for verification. `notice` and
        `access-event.consent` will be null for non-consent records.

        '
      properties:
        response:
          $ref: '#/components/schemas/ResponseMetadata'
        access-record:
          $ref: '#/components/schemas/AccessRecord'
      required:
      - response
      - access-record
    RevokeAccessResponse:
      type: object
      properties:
        response:
          $ref: '#/components/schemas/ResponseMetadata'
        ak:
          $ref: '#/components/schemas/AccessKey'
        revoked-at:
          type: string
          format: date-time
          description: UTC timestamp of revocation.
          example: '2025-06-01T14:22:00Z'
      required:
      - response
      - ak
      - revoked-at
...
    ListAccessRecordsResponse:
      type: object
      properties:
        response:
          $ref: '#/components/schemas/ResponseMetadata'
        mpxn:
          $ref: '#/components/schemas/MPxN'
        access-records:
          type: array
          items:
            $ref: '#/components/schemas/AccessRecordSummary'
      required:
      - response
      - mpxn
      - access-records
...
    ResponseMetadata:
      type: object
      properties:
        resource:
          type: string
          example: /v1/access-records/ak_691df0c788ca043403b7fa90
        timestamp:
          type: string
          format: date-time
          example: '2026-03-11T12:00:00Z'
        transaction-id:
          type: string
          description: Unique identifier for this request/response pair. Quote in
            support enquiries.
          example: tid_691df0c788ca043403b7fa90
      required:
      - resource
      - timestamp
      - transaction-id
...
    ErrorDetail:
      type: object
      properties:
        error-code:
          type: string
          example: VAL001
        message:
          type: string
          example: Field 'mpxn' does not match the required pattern.
      required:
      - error-code
      - message
    ErrorResponse:
      type: object
      properties:
        response:
          $ref: '#/components/schemas/ResponseMetadata'
        errors:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ErrorDetail'
      required:
      - response
      - errors
...